making updates to project structure
