+++ pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   riverpod: ^2.5.1
   flutter_riverpod: ^2.5.1
   go_router: ^14.2.3
+  get_it: ^7.7.0
 
 dev_dependencies:
   flutter_test:
